renamed math_utilities.h to math_utilities.hpp
