+++ applications/GeoMechanicsApplication/custom_constitutive/interface_coulomb_with_tension_cut_off.cpp
@@ -16,7 +16,7 @@
 #include "custom_constitutive/interface_coulomb_with_tension_cut_off.h"
 #include "custom_utilities/check_utilities.hpp"
 #include "custom_utilities/constitutive_law_utilities.h"
-#include "custom_utilities/math_utilities.h"
+#include "custom_utilities/math_utilities.hpp"
 #include "custom_utilities/stress_strain_utilities.h"
 #include "geo_mechanics_application_constants.h"
 #include "geo_mechanics_application_variables.h"
