+++ applications/GeoMechanicsApplication/custom_elements/U_Pw_small_strain_FIC_element.cpp
@@ -15,7 +15,7 @@
 #include "custom_elements/U_Pw_small_strain_FIC_element.hpp"
 #include "custom_utilities/equation_of_motion_utilities.hpp"
 #include "custom_utilities/extrapolation_utilities.h"
-#include "custom_utilities/math_utilities.h"
+#include "custom_utilities/math_utilities.hpp"
 #include "custom_utilities/transport_equation_utilities.hpp"
 
 namespace
