+++ applications/GeoMechanicsApplication/custom_elements/U_Pw_small_strain_element.cpp
@@ -17,7 +17,7 @@
 #include "custom_utilities/constitutive_law_utilities.h"
 #include "custom_utilities/equation_of_motion_utilities.hpp"
 #include "custom_utilities/hydraulic_discharge.h"
-#include "custom_utilities/math_utilities.h"
+#include "custom_utilities/math_utilities.hpp"
 #include "custom_utilities/node_utilities.h"
 #include "custom_utilities/output_utilities.hpp"
 #include "custom_utilities/transport_equation_utilities.hpp"
