+++ applications/GeoMechanicsApplication/custom_utilities/stress_strain_utilities.cpp
@@ -13,7 +13,7 @@
 
 #include "stress_strain_utilities.h"
 #include "custom_utilities/generic_utilities.hpp"
-#include "custom_utilities/math_utilities.h"
+#include "custom_utilities/math_utilities.hpp"
 #include "geo_mechanics_application_constants.h"
 #include "utilities/math_utils.h"
 #include <cmath>
