+++ applications/GeoMechanicsApplication/tests/cpp_tests/custom_utilities/test_Hencky_strain.cpp
@@ -10,7 +10,7 @@
 //  Main authors:    Wijtze Pieter Kikstra
 //
 
-#include "custom_utilities/math_utilities.h"
+#include "custom_utilities/math_utilities.hpp"
 #include "custom_utilities/stress_strain_utilities.h"
 #include "tests/cpp_tests/geo_mechanics_fast_suite.h"
