+++ applications/GeoMechanicsApplication/tests/cpp_tests/custom_utilities/test_stress_strain_utilities.cpp
@@ -11,7 +11,7 @@
 //                   Gennady Markelov
 //
 
-#include "custom_utilities/math_utilities.h"
+#include "custom_utilities/math_utilities.hpp"
 #include "custom_utilities/stress_strain_utilities.h"
 #include "tests/cpp_tests/geo_mechanics_fast_suite.h"
 #include "tests/cpp_tests/test_utilities.h"
